#3873 separate links, update guest home page image deafult, update pnm button, move things to guest view
