+++ src/frontend/src/pages/AdminToolsPage/ProjectsConfig/LinkTypes/LinkTypeTable.tsx
@@ -10,20 +10,22 @@ import NERTable from '../../../../components/NERTable';
 import { isAdmin, LinkType } from 'shared';
 import { useCurrentUser } from '../../../../hooks/users.hooks';
 
-const LinkTypeTable = () => {
+interface LinkTypeTableProps {
+  isOnGuestHomePage?: boolean;
+}
+
+const LinkTypeTable = ({ isOnGuestHomePage }: LinkTypeTableProps) => {
   const currentUser = useCurrentUser();
-  const {
-    data: linkTypes,
-    isLoading: linkTypeIsLoading,
-    isError: linkTypeIsError,
-    error: linkTypeError
-  } = useAllLinkTypes();
+  const { data: links, isLoading: linkTypeIsLoading, isError: linkTypeIsError, error: linkTypeError } = useAllLinkTypes();
   const [createModalShow, setCreateModalShow] = useState<boolean>(false);
   const [showEditModal, setShowEditModal] = useState<boolean>(false);
   const [clickedLinkType, setClickedLinkType] = useState<LinkType>();
 
-  if (!linkTypes || linkTypeIsLoading) return <LoadingIndicator />;
+  if (!links || linkTypeIsLoading) return <LoadingIndicator />;
   if (linkTypeIsError) return <ErrorPage message={linkTypeError.message} />;
+  const linkTypes = links.filter((linkType) =>
+    isOnGuestHomePage ? linkType.isOnGuestHomePage : !linkType.isOnGuestHomePage
+  );
 
   const linkTypeTableRows = linkTypes.map((linkType, index) => (
     <TableRow
@@ -52,7 +54,12 @@ const LinkTypeTable = () => {
 
   return (
     <Box>
-      <CreateLinkTypeModal open={createModalShow} handleClose={() => setCreateModalShow(false)} linkTypes={linkTypes} />
+      <CreateLinkTypeModal
+        open={createModalShow}
+        handleClose={() => setCreateModalShow(false)}
+        linkTypes={linkTypes}
+        isOnGuestHomePage={isOnGuestHomePage}
+      />
       {clickedLinkType && (
         <EditLinkTypeModal
           open={showEditModal}
